build: move all hw maps and info into one central, separate file
Moves everything ddb_regset / ddb_regmap / ddb_info related (which is
used to gather all information required for every piece of supported
hardware) out from ddbridge-main, ddbridge-core and octonet-core into
ddbridge-hw, which then in turn only requires annotations on all
ddb_info structs, and doesn't need to export regsets, regmaps et al.,
thus reducing clutter.
